refactor: migrate to layered architecture (domain/ui/types/utils)
- Move pure business logic (settings, folder-mapping, EagleApiError) to domain/
- Move pure functions (image-format, markdown-image, pseudo-random, etc.) to utils/
- Move Obsidian-dependent code (views, modals, uploader, cache, editor) to ui/
- Keep types/obsidian.d.ts for module augmentation
- ESLint no-restricted-imports enforces domain/types/utils purity
- All 72 tests pass
