+++ scripts/esbuild.config.js
@@ -11,7 +11,7 @@ export const sharedEsbuildConfig = {
   banner: {
     js: banner,
   },
-  entryPoints: ['src/EaglePlugin.ts'],
+  entryPoints: ['src/main.ts'],
   bundle: true,
   external: [
     'obsidian',
