+++ src/domain/EagleApiError.ts
