+++ src/ui/EagleCacheManager.ts
