+++ src/ui/EagleHashStore.ts
