+++ test/eagle-search-items-normalization.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it } from 'vitest'
 
-import { DEFAULT_SETTINGS } from '../src/plugin-settings'
-import EagleUploader from '../src/uploader/EagleUploader'
+import { DEFAULT_SETTINGS } from '../src/domain/settings'
+import EagleUploader from '../src/ui/EagleUploader'
 import { __resetRequestUrlMock, __setRequestUrlMock } from './mocks/obsidian'
 
 function successResponse(data: unknown) {
